Reduce database abstraction layer complexity
This was accomplished by removing trait usage and using simpler
function definitions instead.

A feature flag was also added if more than one database system
has to be supported. The `database` module will export a specific
database system depending on which database feature flag is active.
The accessors of `database` module do not have to know which database
system implementation is in use, as before.