Add `csv` datasource for `Station`
This datasource provides deserialization and database insertion
functionality for the main data source related to bicycle stations.

The `update()` function is generic in terms of `Source`. This means
that all `Sources` that implement the required traits (such as
file or http body readers) can be passed in to this method.