Add csv-based ´Journey` datasource support
This allows inserting journeys to database.

As this often consists of hundreds of thousands of rows, postgres
database insertion performance was manually tested. About 800k rows
can be inserted to a locally running postgres database in about 10 seconds
on a desktop system (AMD 5900x) which is fast enough for this application.