Modify `docker-compose` configuration to allow compiling backend without initialized tables
The problem was that the backend uses the crate `sqlx` which allows compile time
database validation. This requires tables to be created before compilation.

As the application previously only supported automatic table re-generation
during runtime, it was not possible to compile the application without having
the tables initialized.

The solution was to create a Dockerfile for the database and copy
table and view creation related sql files to the container's initdb
folder. These sql files will only be executed if the container
does not have a database created, so these files will usually
only be executed during the first run of `docker-compose up` on
a new system.

The sql files were renamed to lexical order in order for the
postgres container to execute the files in the order that
follows constraints.