Refrain from loading duplicate `Journey`s to the database
The `Journey` csv data source sometimes contained exact duplicate
rows. A decision was made to not insert exact duplicates as this
would confuse users of this application.

`Postgres` database insertion performance is still fine after these
changes, but the insertion logic is somewhat more complicated. As
the `LOAD` function of `Postgres` doesn't support handling `ON CONFLICT`
style statements, a temporary table is created during `Journey` insertion
and afterwards only the new unique rows of this temporary table are inserted
to the main table.

The `Journey` database table definition was also updated to contain
constraint of unique values.