Add reverse proxy configuration, update docker entity naming
Reverse proxy configuration allows the backend and frontend to run on the
"same" address, while being different services.

For example, before reverse proxy configuration, the backend
could listen to the address localhost:3000, but in turn
the frontend service could not bind that address.

Using reverse proxy allows access to the frontend
on localhost:3000, while also allowing access
to the backend service on localhost:3000/api.

Nginx was chosen as the reverse proxy because of relatively
simple configuration and previously good experiences.

----

Docker entities were renamed to be more unique
in order to keep entity naming collisions from happening.