Fix `frontend` Server-Side Rendering (`SSR`) when running through Docker
The frontend service provides server side rendering capability,
allowing the frontend http server to statically render a page
during first load. SSR is useful to allow for example search
engine crawlers to access proper pages, and it can also increase
first page load performance.

The problem was that when the frontend HTTP server was ran
through the docker-compose stack, it tried to access
the previously defined PUBLIC_API_ADDRESS when creating
api requests. As the public api address is not accessible
through the internal docker-compose network, this required
some changes.

(1) A server fetch hook was added to the frontend. This hook
allows replacing an API request's URL when running in the
frontend server instead of the client's web browser.

(2) An additional environment variable, PRIVATE_API_ADDRESS,
was defined in the docker-compose configuration, which was assigned
to be the backend api's internal network address.

(3) Frontend's `Dockerfile` was configured to use this newly defined
PRIVATE_API_ADDRESS environment variable in the container.

SSR support was also configured in a way that continues to allow
a frontend development server to still use the PUBLIC_API_ADDRESS
for SSR, since that is usually available during development.