Run `cargo fmt` and reduce code complexity
Code complexity modifications were mostly variable
assignments. This was done in order to decrease the amount
of new lines that the `cargo fmt` creates, possibly leading
to easierly understandable code in the process.

Some variables were renamed and additionally statement order
was reorganized in a few places. These were minor and there
should not be any logical changes.