Add `.gitattributes` dotfile
The problem was that the backend's `sqlx` crate uses pre-checked
queries when compiling the backend in offline mode. The database
queries are hashed in order to spot differences. Windows often uses
different line endings when compared to for example git (or linux),
so running the `cargo sqlx prepare` command with the originally
saved .sql files on a Windows system lead to the hashes being invalid
when used with a freshly cloned repository.